fix: base job refactor
