+++ src/Common/Common.csproj
@@ -25,6 +25,9 @@
 	  <PackageReference Include="Kralizek.Extensions.Configuration.AWSSecretsManager" Version="1.7.0" />
 	  <PackageReference Include="Microsoft.AspNetCore.Mvc.NewtonsoftJson" Version="6.0.12" />
 	  <PackageReference Include="NReJSON" Version="4.0.0" />
+	  <PackageReference Include="Quartz" Version="3.6.0" />
+	  <PackageReference Include="Quartz.AspNetCore" Version="3.6.0" />
+	  <PackageReference Include="Quartz.Extensions.DependencyInjection" Version="3.6.0" />
 	  <PackageReference Include="RedisGraphDotNet.Client" Version="0.1.0-alpha" />
 	  
       <PackageReference Include="Serilog" Version="2.12.0" />
